requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
Django
django-allauth
idna
oauthlib
pycparser
PyJWT
python-dotenv
requests
requests-oauthlib
sqlparse
tzdata
urllib3
xhtml2pdf

# Google Gemini AI para extracción de contenido (ELIMINADO)
# google-generativeai==0.3.2
PyPDF2==3.0.1

# Dependencias existentes para procesamiento de imágenes
PyMuPDF
pdf2image
Pillow
boto3
django-storages
